pyproject.toml
[project]
name = "managementserver"
version = "0.1.0"
description = "creating the server for management server"
authors = [
    {name = "SUBROTOKOIRY",email = "koirysubroto1103@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask-cors (>=5.0.1,<6.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pymongo (==3.11)",
    "dnspython (>=2.7.0,<3.0.0)",
    "flask-bcrypt (>=1.0.1,<2.0.0)",
    "marshmallow (>=3.26.1,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
